ocr_quick_and_easy/requirements.txt
numpy
matplotlib
pygame
pytest
Pillow
flake8
mypy
pylint
filetype
xxhash
ipython
jupyter
matplotlib-terminal
